small changes and renames
